Create settings file on plugin load and fix debug view to display notes

+++ create_release_zip.ps1
@@ -1,18 +1,31 @@
 Set-Location $PSScriptRoot
 
-$matchInfo = Select-String -Path .\DejaVu.cpp -Pattern 'BAKKESMOD_PLUGIN\(DejaVu, "Deja Vu", "(.*)",'
-$versionString = $matchInfo.Matches.Groups[1].Value
-Write-Output $versionString
+function IsInteractive {
+    # not including `-NonInteractive` since it apparently does nothing
+    # "Does not present an interactive prompt to the user" - no, it does present!
+    $non_interactive = '-command', '-c', '-encodedcommand', '-e', '-ec', '-file', '-f', '-NonInteractive'
+
+    # alternatively `$non_interactive [-contains|-eq] $PSItem`
+    -not ([Environment]::GetCommandLineArgs() | Where-Object -FilterScript {$PSItem -in $non_interactive})
+}
+
+$matchInfo = Select-String -Path .\Version.h -Pattern 'VERSION_.* (.+)'
+$versionString = $matchInfo.Matches.Groups[1].Value + '.' + $matchInfo.Matches.Groups[3].Value + '.' + $matchInfo.Matches.Groups[5].Value + '.' + $matchInfo.Matches.Groups[7].Value + $matchInfo.Matches.Groups[9].Value
+Write-Output "Creating build version: $versionString"
 
 $buildDir = (Get-Location).ToString() + '\zipDir'
-$zipFile = "$buildDir\..\DejaVu - $versionString.zip"
+New-Item -Force -ItemType Directory -Path .\builds | Out-Null
+$zipFile = ".\builds\DejaVu - $versionString.zip"
 
-if (Test-Path $zipFile -PathType Leaf)
-{
+if (Test-Path $zipFile -PathType Leaf) {
+    if (-not (IsInteractive)) {
+        exit
+    }
+
     [System.Reflection.Assembly]::LoadWithPartialName("System.Windows.Forms")
     $msgBoxInput = [System.Windows.Forms.MessageBox]::Show('Release zip already exists. Do you want to overwrite?', 'File exists', 'YesNo', 'Error')
 
-    switch  ($msgBoxInput) {
+    switch ($msgBoxInput) {
         'Yes' {
             Remove-Item $zipFile
         }
@@ -20,11 +33,12 @@ if (Test-Path $zipFile -PathType Leaf)
         'No' {
             exit
         }
-  }
+    }
 }
 
 Remove-Item $buildDir -Recurse
 New-Item -ItemType Directory -Path $buildDir\plugins\settings | Out-Null
+Copy-Item .\x64\Release\DejaVuPlugin.dll -Destination $buildDir\plugins
 Copy-Item dejavu.set -Destination $buildDir\plugins\settings
 New-Item -ItemType Directory -Path $buildDir\source | Out-Null
 Copy-Item .\vendor -Destination $buildDir\source -Recurse
@@ -42,8 +56,7 @@ $fileList = (
     '.\LICENSE',
     '.\RuleSet.ruleset'
 )
-foreach ($file in $fileList)
-{
+foreach ($file in $fileList) {
     Copy-Item $file -Destination $buildDir\source
 }
 
@@ -55,4 +68,6 @@ Add-Type -Assembly System.IO.Compression.FileSystem
     $false
 )
 
-Pause
+if (IsInteractive) {
+    Pause
+}
